requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
openai>=1.10.0

# Vector stores
faiss-cpu>=1.7.4
chromadb>=0.4.22

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.0

# Web framework
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0

# ML/AI
torch>=2.0.0
sentence-transformers>=2.3.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and monitoring
python-json-logger>=2.0.7
